9/1 coding practice

+++ solve_problem/src/level1/practice/practice_nNumbersSpacedByX.java
@@ -4,7 +4,7 @@ package level1.practice;
 public class practice_nNumbersSpacedByX {
 
 	public static void main(String[] args) {
-		int x = 2;
+		long x = 2;
 		int n = 5;
 		
 		for(long l : solution(x,n)) {
@@ -12,10 +12,10 @@ public class practice_nNumbersSpacedByX {
 		}
 
 	}
-    public static long[] solution(int x, int n) {
+    public static long[] solution(long x, int n) {
         long[] answer = new long[n];
-        for(long i=0; i<n; i++) {
-        	answer[(int)i] = x*(i+1);
+        for(int i=0; i<n; i++) {
+        	answer[i] = x*(i+1);
         }
         return answer;
     }
